requirements.txt
Django
gunicorn
django-environ
whitenoise
dj-database-url
psycopg2
djangorestframework